mevbot/core/simd.py
"""SIMD operations for MEV bot using numpy"""
import numpy as np
from typing import List, Tuple

class SIMDProcessor:
    def __init__(self):
        """Initialize SIMD processor"""
        pass
        
    def batch_price_impact(self, amounts: List[float], liquidities: List[float]) -> np.ndarray:
        """Calculate price impact for multiple amounts"""
        amounts = np.array(amounts, dtype=np.float64)
        liquidities = np.array(liquidities, dtype=np.float64)
        
        # Simple constant product AMM formula
        return amounts / (amounts + liquidities)
        
    def batch_profitability(
        self,
        buy_prices: List[float],
        sell_prices: List[float],
        gas_costs: List[float]
    ) -> np.ndarray:
        """Calculate profitability for multiple trades"""
        buy_prices = np.array(buy_prices, dtype=np.float64)
        sell_prices = np.array(sell_prices, dtype=np.float64)
        gas_costs = np.array(gas_costs, dtype=np.float64)
        
        return sell_prices - buy_prices - gas_costs
        
    def batch_sandwich_optimization(
        self,
        target_amounts: List[float],
        pool_liquidities: List[float],
        gas_prices: List[float]
    ) -> Tuple[np.ndarray, np.ndarray]:
        """Optimize sandwich attack amounts for multiple targets"""
        target_amounts = np.array(target_amounts, dtype=np.float64)
        pool_liquidities = np.array(pool_liquidities, dtype=np.float64)
        gas_prices = np.array(gas_prices, dtype=np.float64)
        
        # Calculate optimal front-running amounts (simplified model)
        optimal_amounts = target_amounts * 0.1  # Start with 10% of target amount
        
        # Calculate expected profits
        impact = self.batch_price_impact(optimal_amounts, pool_liquidities)
        profits = target_amounts * impact - gas_prices * 2  # Account for front and back run
        
        return optimal_amounts, profits
